Prevent premature garbage collection of Protocol objects (#8)
# Description

When instantiating a class inheriting the Protocol abstract (WebSocket or any custom protocol), and passing it to the Kuzzle C# constructor, if a reference to the protocol instance is not kept alongside the kuzzle one, then it is garbage collected at the first opportunity.
This triggers a proper call to the C++ Protocol destructor, closing network and memory resources... but the Kuzzle C++ class still uses those, and this ends in a crash because of dangling C++ pointers.

This is because while the Kuzzle class in the C++ layer keeps a pointer to the protocol instance, the corresponding C# Kuzzle class does not. And obviously, the C# garbage collector cannot know that the protocol class instance is still being used in the C++ layer.

This PR configures the SWIG-generated Kuzzle class to force-keep a reference to the protocol instance it receives in a seemingly useless protected class member.